fix: Update data types and formats in the OpenAPI spec (#78)
- Change `format: guid` to `format: uuid` in the OpenAPI spec
- Change `type: number` to `type: integer` in the OpenAPI spec
  (with `format: int64` for the `version` and `totalItems` properties)
- Fix the fallout from these changes and generated code changes

Refs: DSTEW-1400 PDP: Correct OpenAPI `format:guid` and integer field types

+++ provider-data-service/src/main/java/uk/gov/justice/laa/providerdata/controller/ProviderContractManagersController.java
@@ -1,6 +1,5 @@
 package uk.gov.justice.laa.providerdata.controller;
 
-import java.math.BigDecimal;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -33,8 +32,8 @@ public class ProviderContractManagersController implements ContractManagersApi {
       String traceparent,
       List<String> contractManagerId,
       String name,
-      BigDecimal page,
-      BigDecimal pageSize) {
+      Integer page,
+      Integer pageSize) {
 
     var pageParams = PageParamValidator.resolve(page, pageSize);
